Improve Science museum script and fix tests
