+++ tests/dags/providers/provider_api_scripts/resources/sciencemuseum/license_source.json
@@ -1,10 +1,12 @@
 {
-  "legal": {
-    "rights": [
-      {
-        "details": "\u00a9 The Board of Trustees of the Science Museum",
-        "usage_terms": "CC-BY-NC-SA 4.0"
-      }
-    ]
+  "source": {
+    "legal": {
+      "rights": [
+        {
+          "details": "\u00a9 The Board of Trustees of the Science Museum",
+          "usage_terms": "CC-BY-NC-SA 4.0"
+        }
+      ]
+    }
   }
 }
