--- tests/dags/providers/provider_api_scripts/resources/sciencemuseum/no_license.json
@@ -1,7 +1,0 @@
-{
-  "legal": {
-    "rights": [
-      {}
-    ]
-  }
-}
