+++ tests/dags/providers/provider_api_scripts/resources/sciencemuseum/no_license_no_legal.json
@@ -1,0 +1,10 @@
+{
+  "source": {
+    "title": [
+      {
+        "type": "caption",
+        "value": "Glass measuring cylinder from Luke Howard's rain-gauge"
+      }
+    ]
+  }
+}
