+++ tests/dags/providers/provider_api_scripts/resources/sciencemuseum/no_license_no_usage_terms.json
@@ -1,0 +1,11 @@
+{
+  "source": {
+    "legal": {
+      "rights": [
+        {
+          "details": "\u00a9 The Board of Trustees of the Science Museum, London"
+        }
+      ]
+    }
+  }
+}
